frontend: refactor other routes
