+++ apps/frontend/src/components/wydarzenia/ListaWydarzenKonia.tsx
