+++ apps/frontend/src/components/wydarzenia/Profilaktyczne.tsx
@@ -1,4 +1,4 @@
-import HorseEventList from "../konie/ListaZdarzenKonia";
+import HorseEventList from "./ListaWydarzenKonia";
 function Profilaktyczne() {
   return <HorseEventList type="zdarzenia_profilaktyczne" />;
 }
